pyproject.toml
[project]
name = "killogram-messenger"
version = "0.1.0"
description = "FastAPI Messenger with JWT and PostgreSQL"
authors = [
    {name = "Eugene Pisarev", email = "zm1s3ry@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt (==4.0.1)",
    "python-multipart>=0.0.10", # Нужен для работы OAuth2PasswordRequestForm
    "redis>=5.2.0",
    "orjson>=3.10.0",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0"
]

[tool.poetry]
# В Poetry 2.0+ основные метаданные берутся из [project]
# Здесь можно оставить настройки виртуального окружения
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"